Lab I2 complete

+++ danilenko.ekaterina/I2/MaxElements.cpp
@@ -1,10 +1,10 @@
 #include "MaxElements.h"
 
-int CountMaxElements(const int *array, const size_t size)
+int CountMaxElements(const int *array, const int size)
 {
-  size_t count = 0;
+  int count = 0;
   int max_val = 0;
-  for (size_t i = 0; i < size; i++)
+  for (int i = 0; i < size; i++)
     if (array[i] > max_val)
     {
       max_val = array[i];
