+++ danilenko.ekaterina/I2/PositiveRepeatingNumber.cpp
@@ -1,9 +1,9 @@
 #include "PositiveRepeatingNumber.h"
 
-int PosRepeatingNum(const int *array, const size_t size)
+int PosRepeatingNum(const int *array, const int size)
 {
-  size_t count = 0;
-  for (size_t i = 0; i < size - 1; i++)
+  int count = 0;
+  for (int i = 0; i < size - 1; i++)
   {
     if (array[i] == array[i - 1])
       count++;
