+++ danilenko.ekaterina/I2/main.cpp
@@ -13,7 +13,7 @@ int main(int argc, char *array[])
   }
 
   //1
-  size_t size1 = 10;
+  int size1 = 10;
   int array1[10] = {1, 9, 8, 4, 9, 9, 9, 4 ,3, 2};
   try
   {
@@ -28,14 +28,14 @@ int main(int argc, char *array[])
   }
 
   //2
-  size_t size2 = 0;
+  int size2 = 0;
   int* array2 = 0;
   std::cin >> size2;
   if (size2 > 0)
   {
     array2 = new int[size2];
     std::srand(time(nullptr));
-    for (size_t i = 0; i < size2; i++)
+    for (int i = 0; i < size2; i++)
     {
       array2[i] = rand()%10;
     }
@@ -61,14 +61,14 @@ int main(int argc, char *array[])
   if (file.is_open())
   {
     int* array3;
-    size_t size3 = 0;
+    int size3 = 0;
     file >> size3;
     if (file)
     {
       if (size3 > 0)
       {
         array3 = new int[size3];
-        for (size_t i = 0; i < size3; i++)
+        for (int i = 0; i < size3; i++)
         {
           file >> array3[i];
           if (!file)
